fix: the VC-1 level printed as a control character, not a number
The level is a uint8_t, and streaming one of those prints the character
with that code. Advanced profile level 3 came out as

  Profile: Advanced@  (followed by the byte 0x03)

which is issue 883, filed as a cosmetic bug with a screenshot showing a
heart where the level should be, because that is what code 3 looks like in
the console font the reporter was using.

  before   Profile: Advanced@^C Resolution: 1920:1080p
  after    Profile: Advanced@3  Resolution: 1920:1080p

Reproduced on two VC-1 files here, so this needed nothing from the reporter.

+++ tsMuxer/vc1Parser.cpp
@@ -77,7 +77,10 @@ string VC1SequenceHeader::getStreamDescr() const
         rez << "Complex";
         break;
     case Profile::ADVANCED:
-        rez << "Advanced@" << level;
+        // level is a uint8_t, and streaming one prints the CHARACTER with that code rather than
+        // the number. Advanced profile level 3 came out as "Advanced@" followed by a control
+        // character, which is what issue 883 reports as a cosmetic bug.
+        rez << "Advanced@" << static_cast<int>(level);
         break;
     }
